requirements.txt
beautifulsoup4==4.12.0
bs4==0.0.1
pyTelegramBotAPI==4.10.0
requests==2.28.2
schedule==1.1.0
telebot==0.0.5
python_dotenv==1.0.0